requirements.txt
celery==5.0.5
Django==3.1.5
django-cors-headers==3.6.0
django-model-utils==4.1.1
django-polymorphic==3.0.0
djangorestframework==3.12.2
gunicorn==20.0.4
ipython==7.21.0
psycopg2-binary==2.8.6
requests==2.25.1
martor==1.5.8
django-decouple==2.1
Pillow==8.1.0
python-telegram-bot==13.2
django-allauth==0.44.0
google-auth==1.26.1
djangorestframework-simplejwt==4.6.0
dj-rest-auth==2.1.3
django-import-export==2.5.0
zeep==4.0.0
drf-api-tracking==1.7.8
PySocks==1.7.1
